Enable trampolining of async state execution
This replaces the previous body of an `async` block with a while
loop:

    var task$async: Option[() => Unit] = Some(<future body>)

    while (task$async.nonEmpty) {
      val task = task$async.get
      task$async = None
      task()
    }

This enables executing an `async` block on a single thread by using
a future system that blocks upon `await`. This execution mode can be
very useful for debugging. The introduced trampolining enables this
mode without running into stack overflows.