Fix crashers in do/while and while(await(..))
The new tree shapes handled for do/while look like:

    // type checked
    async({
      val b = false;
      doWhile$1(){
        await(());
        if (b)
          doWhile$1()
        else
          ()
      };
      ()
    })

We had to change ExprBuilder to create states for the if/else
that concludes the doWhile body, and also loosen the assertion
that the label jump must be the last thing we see.

We also have to look for more than just `containsAwait` when
deciding whether an `If` needs to be transformed into states;
it might also contain a jump to the enclosing label that is on
the other side of an `await`, and hence needs to be a state
transition instead.