Avoid compiler warning when awaiting Future[Unit]
During the ANF transform, we were generating a tree of the shape:

   {
     val temp: Unit = await(futureOfUnit)
     temp
     ()
   }

I tried to simplifiy this to avoid creating the temporary value,
but this proved difficult as it would have required changes to
the subsequent state machine transformation.

Even replacing `temp` with `()` made the state machine transform
harder.

So for now, I've just inserted `temp.asInstanceOf[Unit]` to hide
from the compiler warning.

Fixes #74