Avoid leaking untyped trees out of macro
The stack trace and bisection in #119 made me notice that we
are failing to typecheck the cast we generated in the fix for #74.

The ticket doesn't have a reproduction, so I'm submitting this
without a test case.

Fixes #119

(cherry picked from commit a037da192dbea32cffa8c5c8d33e7c56af0c353f)